requirements/base_updated.txt
# Core dependencies - Updated to latest stable versions
pytz==2025.2  # https://github.com/stub42/pytz
Pillow==11.3.0  # https://github.com/python-pillow/Pillow        
whitenoise==6.9.0  # https://github.com/evansd/whitenoise

# Django - Latest LTS version
# ------------------------------------------------------------------------------
django==5.0.11  # Latest LTS version # https://www.djangoproject.com/
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-filter==25.1 # https://github.com/carltongibson/django-filter
django-modeltranslation==0.19.16 # https://github.com/Buren/django-modeltranslation

# PDF generator
reportlab==4.4.3
xhtml2pdf==0.2.17

# Customize django admin - Updated version
django-jet-reboot==1.3.10

# Environment variable
python-decouple==3.8

# Payments
stripe==12.4.0
gopay==2.2.5